Initial commit of project
This is a reproduction of the project that my team works in.
In Windows, we can rename folders to be a different case. SBT and Metals don't seem to have an issue compiling but we can't debug or scalafix folders with a naming convention that doesn't match what is in the build.sbt file.